Fetch all init data with one gql request (#827)
Fetch cluster microfrontends via graphql
